Refactors to make app compatible with google3 appengine.
