+++ repl/appengine/web/src/app.ts
@@ -3,8 +3,6 @@
  */
 
 
-let replApp =
-
  (function() {
 
   interface Stringer {
@@ -35,81 +33,76 @@ let replApp =
         .replace(/>/g, '&gt;');
   }
 
-  class AppImpl {
-    showError(error: Stringer) {
-      document.querySelector('pre#error-box')
-          .innerHTML = quoteattr(error.toString())
-    }
-
-    render (request, response) {
-      let insertionPoint = document.querySelector('div.input-block');
-      let html = '';
-      let lastResp = '';
-    
-      request.commands.forEach(function(reqLine: String, i: number) {
-        let resp : any = null;
-        if (i < response.responses.length) {
-          resp = response.responses[i];
-        }
-        console.log(response, resp);
-        let error = !!resp.issue;
-        let respLine = error ? resp.issue || 'err' : resp.replOutput || 'ok';
-        lastResp = respLine;
-        let preview = trimLen(respLine);
-    
-        html += `<div class="statement-block">
-            &gt; <input class="repl-stmt" value="${quoteattr(reqLine)}"><br>
-            <code class="${error ? 'repl-out-error' : 'repl-out-ok'}">${
-            quoteattr(preview)}</code>
-          </div>`;
-      });
-    
-      html += newInput;
-    
-      insertionPoint.innerHTML = html;
-    
-      document.querySelector('code#result').innerHTML = quoteattr(lastResp);
-      this.readyForInput();
-
-    }
-
-    submit() {
-      let inputs = document.querySelectorAll('input.repl-stmt');
-      let req = {'commands': []};
-      inputs.forEach(function(el) {
-        let inp = <HTMLInputElement>el;
-        if (inp.value != null && inp.value.trim() != '') {
-          req.commands.push(inp.value);
-        }
-      });
-      let payload = JSON.stringify(req);
-      console.log('submitting: %s', payload);
-      window
-          .fetch(endpoint, {
-            method: 'POST',
-            headers: {'Content-Type': 'application/json'},
-            body: payload
-          })
-          .then(response => response.json())
-          .then(data => this.render(req, data))
-          .catch(ex => this.showError(ex));
-    }
-
-    readyForInput() {
-      let mainInput = <HTMLInputElement>document.querySelector('input.repl-stmt-new');
-      let submitCallback = () => this.submit();
-      mainInput.focus();
-
-      mainInput.addEventListener('keyup', function(e : KeyboardEvent) {
-        if (e.code == 'Enter' && !e.metaKey && !e.ctrlKey) {
-          e.preventDefault();
-          submitCallback();
-        }
-      });
-    }
+  function showError(error: Stringer) {
+    document.querySelector('pre#error-box')
+        .innerHTML = quoteattr(error.toString())
   }
 
-  let app = new AppImpl;
+  function render (request, response) {
+    let insertionPoint = document.querySelector('div.input-block');
+    let html = '';
+    let lastResp = '';
+  
+    request.commands.forEach(function(reqLine: String, i: number) {
+      let resp : any = null;
+      if (i < response.responses.length) {
+        resp = response.responses[i];
+      }
+      console.log(response, resp);
+      let error = !!resp.issue;
+      let respLine = error ? resp.issue || 'err' : resp.replOutput || 'ok';
+      lastResp = respLine;
+      let preview = trimLen(respLine);
+  
+      html += `<div class="statement-block">
+          &gt; <input class="repl-stmt" value="${quoteattr(reqLine)}"><br>
+          <code class="${error ? 'repl-out-error' : 'repl-out-ok'}">${
+          quoteattr(preview)}</code>
+        </div>`;
+    });
+  
+    html += newInput;
+  
+    insertionPoint.innerHTML = html;
+  
+    document.querySelector('code#result').innerHTML = quoteattr(lastResp);
+    readyForInput();
+
+  }
+
+  function submit() {
+    let inputs = document.querySelectorAll('input.repl-stmt');
+    let req = {'commands': []};
+    inputs.forEach(function(el) {
+      let inp = <HTMLInputElement>el;
+      if (inp.value != null && inp.value.trim() != '') {
+        req.commands.push(inp.value);
+      }
+    });
+    let payload = JSON.stringify(req);
+    console.log('submitting: %s', payload);
+    window
+        .fetch(endpoint, {
+          method: 'POST',
+          headers: {'Content-Type': 'application/json'},
+          body: payload
+        })
+        .then(response => response.json())
+        .then(data => render(req, data))
+        .catch(ex => showError(ex));
+  }
+
+  function readyForInput() {
+    let mainInput = <HTMLInputElement>document.querySelector('input.repl-stmt-new');
+    mainInput.focus();
+
+    mainInput.addEventListener('keyup', function(e : KeyboardEvent) {
+      if (e.code == 'Enter' && !e.metaKey && !e.ctrlKey) {
+        e.preventDefault();
+        submit();
+      }
+    });
+  }
   
   window.addEventListener('load', () => {
     let evalBtn = document.querySelector('button#evaluate');
@@ -120,9 +113,8 @@ let replApp =
     }
     let insertionPoint = document.querySelector('div.input-block');
     insertionPoint.innerHTML = newInput;
-    evalBtn.addEventListener('click', () => app.submit());
-    app.readyForInput();
+    evalBtn.addEventListener('click', () => submit());
+    readyForInput();
   });
 
-  return app;
  })();
